Authorization fixed. Several logic and functionalities added.

+++ BookyBook.API/Controllers/UserController.cs
@@ -1,12 +1,12 @@
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Identity;
-using BookyBook.Data;
 using BookyBook.Business;
 using BookyBook.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace API.Controllers;
 
 [ApiController]
+[Authorize]
 [Route("[controller]")]
 public class UserController : ControllerBase
 {
@@ -20,37 +20,44 @@ public class UserController : ControllerBase
         _userService = userService;
     }
 
+    [Authorize(Roles = Roles.Admin)]
     [HttpGet(Name = "GetUsers")]
-    public ActionResult<IEnumerable<User>> GetUsers([FromQuery] UserQueryParameters userQueryParameters)
+    public ActionResult<IEnumerable<UserLogedDTO>> GetUsers([FromQuery] UserQueryParameters userQueryParameters, [FromQuery] string? sortBy)
     {
         if (!ModelState.IsValid)  {return BadRequest(ModelState); } 
         try 
         {
-            var users = _userService.GetAllUsers(userQueryParameters);
+            if (sortBy == null) {sortBy = "";}
+            var users = _userService.GetAllUsers(userQueryParameters, sortBy);
             return Ok(users);
         }     
         catch (Exception ex)
         {
-            _logger.LogInformation(ex.Message);
-            return BadRequest("No se han encontrado usuarios");
+            _logger.LogInformation(ex.ToString());
+            return BadRequest(ex.Message);
         }
     }
 
+    
     [HttpGet("{userId}/borrowings", Name = "GetBorrowingsByUserId")]
-    public IActionResult GetBorrowingsByUserId(int userId, [FromQuery] UserQueryParameters userQueryParameters)
+    public IActionResult GetBorrowingsByUserId(int userId, [FromQuery] UserQueryParameters userQueryParameters, [FromQuery] string? sortBy)
     {
         if (!ModelState.IsValid)  {return BadRequest(ModelState); } 
+        if (!_userService.HasAccessToResource(userId, HttpContext.User)) 
+            {return Forbid(); }
         try
         {
-            var borrowings = _userService.GetBorrowingsByUserId(userId, userQueryParameters);
+            if (sortBy == null) {sortBy = "";}
+            var borrowings = _userService.GetBorrowingsByUserId(userId, userQueryParameters, sortBy);
             return Ok(borrowings);
         }
         catch (Exception ex)
         {
-            _logger.LogInformation(ex.Message);
-            return BadRequest("No se han encontrado préstamos");
+            _logger.LogInformation(ex.ToString());
+            return BadRequest(ex.Message);
         }
     }
+
 
     [HttpGet("{userId}", Name = "GetUser")]
     public IActionResult GetUser(int userId)
@@ -64,23 +71,24 @@ public class UserController : ControllerBase
         }
         catch (KeyNotFoundException ex)
         {
-            _logger.LogInformation(ex.Message);
-           return NotFound("No encontrado el usuario " + userId);
+            _logger.LogInformation(ex.ToString());
+           return NotFound(ex.Message);
         }
     }
 
+    [Authorize(Roles = Roles.Admin)]
     [HttpPost()]
-    public IActionResult CreateUser([FromBody] UserDtoIn userDtoIn)
+    public IActionResult CreateUser([FromBody] UserCreateDTO userCreateDTO)
     {
-        if (!ModelState.IsValid)  {return BadRequest(ModelState); } 
+        if (!ModelState.IsValid)  {return BadRequest(ModelState); }
         try {
-            _userService.AddUser(userDtoIn);
-            return Ok(userDtoIn);
+            _userService.AddUser(userCreateDTO);
+            return Ok(userCreateDTO);
         }     
         catch (Exception ex)
         {
-            _logger.LogInformation(ex.Message);
-            return BadRequest("No se ha podido crear el usuario");
+            _logger.LogInformation(ex.ToString());
+            return BadRequest(ex.Message);
         }
         
     }
@@ -88,16 +96,17 @@ public class UserController : ControllerBase
     [HttpPut("{userId}")]
     public IActionResult UpdateUser(int userId, [FromBody] UserUpdateDTO userUpdate)
     {
-        if (!ModelState.IsValid)  {return BadRequest(ModelState); } 
+        if (!ModelState.IsValid)  {return BadRequest(ModelState); }
+        if (!_userService.HasAccessToResource(userId, HttpContext.User)) 
+            {return Forbid(); }
         try
         {
             _userService.UpdateUser(userId, userUpdate);
-            //return NoContent();
             return Ok(_userService.GetUser(userId));
         }
         catch (KeyNotFoundException ex)
         {
-            _logger.LogInformation(ex.Message);
+            _logger.LogInformation(ex.ToString());
             return NotFound("No encontrado el usuario " + userId);
         }
     }
@@ -105,15 +114,16 @@ public class UserController : ControllerBase
     [HttpDelete("{userId}")]
     public IActionResult DeleteUser(int userId)
     {
+        if (!_userService.HasAccessToResource(userId, HttpContext.User)) 
+            {return Forbid(); }
         try
         {
             _userService.DeleteUser(userId);
-            //return NoContent();
             return Ok(_userService.GetUser(userId));
         }
         catch (KeyNotFoundException ex)
         {
-            _logger.LogInformation(ex.Message);
+            _logger.LogInformation(ex.ToString());
             return NotFound("No encontrado el usuario " + userId);
         }
     }
